+++ examples/demo-module/admin/demo_admin.php
@@ -2,17 +2,20 @@
 
 declare(strict_types=1);
 
+use Bitrix\Main\Loader;
 use MB\Bitrix\AdminKit\Manager\AdminKitManager;
 use MB\Bitrix\AdminKit\Manager\AdminKitScope;
 
 require $_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/main/include/prolog_admin_before.php';
+
+Loader::includeModule('demo.module');
 
 global $APPLICATION, $adminPage;
 $adminPage->hideTitle();
 
 require $_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/main/include/prolog_admin_after.php';
 
-$scope = new AdminKitScope('demo.module', ['local/modules/demo.module/lib/Admin']);
+$scope = AdminKitScope::fromModuleId('demo.module');
 (new AdminKitManager($scope))->getCurrentPage()->render();
 
 require $_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/main/include/epilog_admin.php';